# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Señoriales API",
    "version": "1.0.0",
    "description": "API del backend monolito de Señoriales. Cubre auth, sesiones de práctica, escenarios, integraciones de ElevenLabs y endpoints administrativos."
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Desarrollo local"
    },
    {
      "url": "https://centro-de-negocios.org",
      "description": "Producción"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "PracticeSession": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "userId": {
            "type": "string",
            "format": "uuid"
          },
          "scenarioId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "durationSeconds": {
            "type": "integer"
          },
          "score": {
            "type": "integer",
            "nullable": true
          },
          "passed": {
            "type": "boolean"
          },
          "rating": {
            "type": "integer",
            "nullable": true
          },
          "aiFeedback": {
            "type": "string",
            "nullable": true
          },
          "connectMs": {
            "type": "integer",
            "nullable": true
          },
          "ttfaSamplesMs": {
            "type": "array",
            "items": {
              "type": "integer"
            }
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "EvaluationResult": {
        "type": "object",
        "properties": {
          "score": {
            "type": "integer",
            "minimum": 0,
            "maximum": 100
          },
          "passed": {
            "type": "boolean"
          },
          "feedback": {
            "type": "string"
          },
          "breakdown": {
            "type": "object",
            "properties": {
              "apertura": {
                "type": "integer"
              },
              "escucha_activa": {
                "type": "integer"
              },
              "manejo_objeciones": {
                "type": "integer"
              },
              "propuesta_valor": {
                "type": "integer"
              },
              "cierre": {
                "type": "integer"
              }
            }
          }
        }
      },
      "Scenario": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "objection": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "difficulty": {
            "type": "string",
            "enum": [
              "easy",
              "medium",
              "hard"
            ]
          },
          "voiceType": {
            "type": "string",
            "enum": [
              "male",
              "female"
            ]
          },
          "firstMessage": {
            "type": "string"
          },
          "displayOrder": {
            "type": "integer"
          }
        }
      },
...
      "UserProgress": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "userId": {
            "type": "string",
            "format": "uuid"
          },
          "scenarioId": {
            "type": "string",
            "format": "uuid"
          },
          "isUnlocked": {
            "type": "boolean"
          },
          "isCompleted": {
            "type": "boolean"
          },
          "bestScore": {
            "type": "integer",
            "nullable": true
          },
          "attempts": {
            "type": "integer"
          },
          "firstCompletedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "lastAttemptAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/api/scenarios/{id}": {
      "get": {
        "tags": [
          "Scenarios"
        ],
        "summary": "Devuelve un escenario por id",
        "security": [],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Escenario",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Scenario"
                }
              }
            }
          },
          "404": {
            "description": "No encontrado"
          }
        }
      }
    },
...
    "/api/sessions": {
      "get": {
        "tags": [
          "Sessions"
        ],
        "summary": "Lista las sesiones del usuario autenticado",
        "responses": {
          "200": {
            "description": "Lista de sesiones",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PracticeSession"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Sessions"
        ],
        "summary": "Crea una sesión nueva (al iniciar una llamada)",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "scenarioId": {
                    "type": "string",
                    "nullable": true
                  },
                  "durationSeconds": {
                    "type": "integer",
                    "default": 0
                  },
                  "abVariantId": {
                    "type": "string",
                    "nullable": true
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Sesión creada",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PracticeSession"
                }
              }
            }
          }
        }
      }
    },
    "/api/sessions/{id}": {
      "patch": {
        "tags": [
          "Sessions"
        ],
        "summary": "Actualiza una sesión (duración, score, latencia, etc.)",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "durationSeconds": {
                    "type": "integer"
                  },
                  "score": {
                    "type": "integer"
                  },
                  "passed": {
                    "type": "boolean"
                  },
                  "rating": {
                    "type": "integer",
                    "minimum": 1,
                    "maximum": 5
                  },
                  "aiFeedback": {
                    "type": "string"
                  },
                  "connectMs": {
                    "type": "integer",
                    "nullable": true
                  },
                  "ttfaSamplesMs": {
                    "type": "array",
                    "items": {
                      "type": "integer"
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Sesión actualizada",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PracticeSession"
                }
              }
            }
          },
          "404": {
            "description": "Sesión no encontrada o no pertenece al usuario"
          }
        }
      }
    },
    "/api/sessions/evaluate": {
      "post": {
        "tags": [
          "Sessions"
        ],
        "summary": "Evalúa el transcript de una sesión y persiste el puntaje",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "transcript"
                ],
                "properties": {
                  "sessionId": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "transcript": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "role": {
                          "type": "string",
                          "enum": [
                            "user",
                            "agent"
                          ]
                        },
                        "content": {
                          "type": "string"
                        },
                        "timestamp": {
                          "type": "integer"
                        }
                      }
                    }
                  },
                  "scenarioId": {
                    "type": "string",
                    "format": "uuid",
                    "nullable": true
                  },
                  "durationSeconds": {
                    "type": "integer"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Resultado de la evaluación",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EvaluationResult"
                }
              }
            }
          }
        }
      }
    },
    "/api/sessions/{id}/transcript": {
      "post": {
        "tags": [
          "Sessions"
        ],
        "summary": "Guarda el transcript completo de la sesión",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "transcript": {
                    "type": "array",
                    "items": {
                      "type": "object"
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "get": {
        "tags": [
          "Sessions"
        ],
        "summary": "Obtiene el transcript de la sesión",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Transcript"
          },
          "404": {
            "description": "Sesión no encontrada"
          }
        }
      }
    },
    "/api/progress": {
      "get": {
        "tags": [
          "Progress"
        ],
        "summary": "Devuelve el progreso por escenario del usuario autenticado",
        "responses": {
          "200": {
            "description": "Lista de progreso por escenario (ordenada por displayOrder del escenario)",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UserProgress"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Progress"
        ],
        "summary": "Crea o actualiza (upsert) el progreso del usuario en un escenario",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "scenarioId"
                ],
                "properties": {
                  "scenarioId": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "isUnlocked": {
                    "type": "boolean"
                  },
                  "isCompleted": {
                    "type": "boolean"
                  },
                  "bestScore": {
                    "type": "integer"
                  },
                  "attempts": {
                    "type": "integer"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Progreso actualizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserProgress"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/users/me/unlock-level2": {
      "post": {
        "tags": [
          "Users"
        ],
        "summary": "Desbloquea Manejo de Objeciones (Nivel 2) para el usuario actual",
        "description": "Requiere que el usuario tenga al menos una sesión de examen final aprobada (sin scenarioId).",
        "responses": {
          "200": {
            "description": "Nivel 2 desbloqueado",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "level2Unlocked": {
                      "type": "boolean"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Examen final no aprobado"
          }
        }
      }
    },
...
  }
}